+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for unicode characters in Python 3.8.
 - Added `--vendor PATH` command to download the program for offline use.
+- Updated `--quiet` output to include actual program versions.
 
 # 3.0.2 (2020-01-06)